Added flag to configure.ac for verbose-debug-output
Added `-enable-verbose` flag to configure.ac to allow for the output of debug information. This allows for accessing information such as gradients or virial, which are required for implementing tests.

I have done some simple testing, and it does appear that `./configure -enable-verbose` does cause the `-DDEBUG` flag to be correctly passed to the compiler.